SFDM-3384 [Back][CX] Доработать метод PATCH /api​/cx​/v1​/bpmn/cj/{cj-id} заполнения атрибутов CJ корректной сортировкой

+++ src/main/java/ru/beeline/cxbackend/dto/CjStepDto.java
@@ -6,9 +6,11 @@ package ru.beeline.cxbackend.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class CjStepDto {
     private String name;
-    private Integer order;
+    private BigDecimal order;
     private String description;
 }
